requirements.txt
livekit-agents
livekit-plugins-openai
livekit-plugins-silero
livekit-plugins-google
livekit-plugins-noise-cancellation
duckduckgo-search
langchain-community
requests
python-dotenv
duckduckgo-search
langchain-community
requests
python-dotenv